Update Copyright and start work on Relocated version of Reversion
