+++ .idea/copyright/profiles_settings.xml
@@ -1,3 +1,3 @@
 <component name="CopyrightManager">
-  <settings default="GLP3" />
+  <settings default="MIT" />
 </component>
