+++ geyser-reversion/pom.xml
@@ -1,20 +1,26 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
-  ~ EduSupport - Minecraft Protocol Support for MultiVersion in Geyser
-  ~ Copyright (C) 2020 GeyserReversion Developers
+  ~ MIT License
   ~
-  ~ This program is free software: you can redistribute it and/or modify
-  ~ it under the terms of the GNU General Public License as published by
-  ~ the Free Software Foundation, either version 3 of the License, or
-  ~ (at your option) any later version.
+  ~ Copyright (c) 2020 GeyserReversion Developers
   ~
-  ~ This program is distributed in the hope that it will be useful,
-  ~ but WITHOUT ANY WARRANTY; without even the implied warranty of
-  ~ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  ~ GNU General Public License for more details.
+  ~ Permission is hereby granted, free of charge, to any person obtaining a copy
+  ~ of this software and associated documentation files (the "Software"), to deal
+  ~ in the Software without restriction, including without limitation the rights
+  ~ to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+  ~ copies of the Software, and to permit persons to whom the Software is
+  ~ furnished to do so, subject to the following conditions:
   ~
-  ~ You should have received a copy of the GNU General Public License
-  ~ along with this program.  If not, see <http://www.gnu.org/licenses/>.
+  ~ The above copyright notice and this permission notice shall be included in all
+  ~ copies or substantial portions of the Software.
+  ~
+  ~ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+  ~ IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+  ~ FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+  ~ AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+  ~ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+  ~ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+  ~ SOFTWARE.
   -->
 
 <project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
@@ -37,46 +43,9 @@
             <version>1.2.0-SNAPSHOT</version>
             <scope>provided</scope>
         </dependency>
-        <!-- Translators -->
         <dependency>
             <groupId>au.com.grieve.reversion</groupId>
-            <artifactId>translator-v390ee_to_v408be</artifactId>
-            <version>${reversion.version}</version>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
-            <groupId>au.com.grieve.reversion</groupId>
-            <artifactId>translator-v409be_to_v408be</artifactId>
-            <version>${reversion.version}</version>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
-            <groupId>au.com.grieve.reversion</groupId>
-            <artifactId>translator-v411be_to_v409be</artifactId>
-            <version>${reversion.version}</version>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
-            <groupId>au.com.grieve.reversion</groupId>
-            <artifactId>translator-v412be_to_v411be</artifactId>
-            <version>${reversion.version}</version>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
-            <groupId>au.com.grieve.reversion</groupId>
-            <artifactId>translator-v414be_to_v412be</artifactId>
-            <version>${reversion.version}</version>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
-            <groupId>au.com.grieve.reversion</groupId>
-            <artifactId>edition-education</artifactId>
-            <version>${reversion.version}</version>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
-            <groupId>au.com.grieve.reversion</groupId>
-            <artifactId>edition-bedrock</artifactId>
+            <artifactId>build-all-relocated</artifactId>
             <version>${reversion.version}</version>
             <scope>compile</scope>
         </dependency>
@@ -100,16 +69,8 @@
                             <artifactSet>
                                 <includes>
                                     <include>au.com.grieve.reversion:*</include>
-                                    <include>com.nukkitx.protocol:*</include>
                                 </includes>
                             </artifactSet>
-                            <relocations>
-                                <relocation>
-                                    <pattern>com.nukkitx.protocol</pattern>
-                                    <shadedPattern>au.com.grieve.geyser.reversion.shaded.nukkitx.protocol
-                                    </shadedPattern>
-                                </relocation>
-                            </relocations>
                         </configuration>
                     </execution>
                 </executions>
